+++ src/api/mod.rs
@@ -1,14 +1,7 @@
 // mod.rs : src/api
 
-macro_rules! declare_and_publish {
-    ($mod_name:ident, $($type_name:ident),* $(,)?) => {
-        mod $mod_name;
+use crate::macros::declare_and_publish;
 
-        pub use $mod_name::{
-            $($type_name),*
-        };
-    };
-}
 
 declare_and_publish!(evaluate_scalar, evaluate_scalar_eq_approx);
 declare_and_publish!(evaluate_vector, evaluate_vector_eq_approx);