+++ src/traits/mod.rs
@@ -2,15 +2,8 @@
 
 //! Traits.
 
-macro_rules! declare_and_publish {
-    ($mod_name:ident, $($type_name:ident),* $(,)?) => {
-        mod $mod_name;
+use crate::macros::declare_and_publish;
 
-        pub use $mod_name::{
-            $($type_name),*
-        };
-    };
-}
 
 declare_and_publish!(approximate_equality_evaluator, ApproximateEqualityEvaluator);
 declare_and_publish!(testable_as_f64, TestableAsF64);